Turret Build
This push was focused on being able to build turrets on top of nodes. These turrets will work straight away from being built and will have an off set that will place them centre placed on a node.

Nodes where turrets can be built will change colour if the node does not contain an already constructed turret. If the player tries to place a turret on top of a built turret then they will be greeted with a message informing them that they cannot.