Camera Control
This Contains a script that allows the player to control the camera and move it around the level either by using the WASD keys or by placing the mouse on the corner of the game screen.

If the player presses ESC then they will be taken out of camera control and will have free use of the mouse again.

The player is able to zoom in an out of the scene with the scroll wheel allowing them to get a closer look, these values are set by a public min max variable.